Lamp + Syndie Omnitool Rebrand + Beauty Detector Sprites <!-- The text between the arrows are comments - they will not be visible on your PR. --> <!-- To automatically tag this PR, add the uppercase label(s) surrounded by brackets below, for example: [LABEL] -->
## About the PR <!-- Describe the Pull Request here. What does it change? What other things could this impact? -->
Contains a resprite of the lamp, a reflavor/sprite of the Syndie omnitool, and the beauty detector sprite that I never marged.

## Why's this needed? <!-- Describe why you think this should be added to the game. -->

Shrugs

## Changelog
<!-- If necessary, put your changelog entry below. Otherwise, please delete it.
Use however you want to be credited in the changelog in place of CodeDude.
Use (*) for major changes and (+) for minor changes. For example: -->
(+) Updated Lamp Sprite
(+) Reflavored Syndie Omnitool into FOSS Omnitool.
(+) Tosses in the beauty_detector sprite
```changelog
(u)CodeDude
(*)Added soft soft pizza to the game - a tasty drink found in soda machines!
(+)Made some small sprite changes to the base pizza sprite.
```